Aboneler api kısmı ayarlanması gerceklestirild,

+++ app/Http/Controllers/api/back/aboneler/indexController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\api\back\aboneler;
 
+use App\Helpers\myHelper;
 use App\Http\Controllers\Controller;
 use App\Models\AboneModel;
 use Illuminate\Http\Request;
@@ -9,7 +10,30 @@ use Yajra\DataTables\DataTables;
 
 class indexController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (myHelper::yetkiKontrol('aboneler',"aktiflik")===false){
+                return redirect()->route('back.home.index')->with(array(
+                    "type" => "error",
+                    "title" => "Hata",
+                    "text" => "Yetkiniz Yok"
+                ));
+            }
+            return $next($request);
+        });
+    }
+
     public function index(){
+
+        if (myHelper::yetkiKontrol('aboneler',"listeleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $query = AboneModel::query();
         $data = DataTables::of($query)
             ->addIndexColumn()
@@ -35,6 +59,15 @@ class indexController extends Controller
 
     // EKLEME KISMI
     public function store(Request $request){
+
+        if (myHelper::yetkiKontrol('aboneler',"ekleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = $request->except("_token");
 
         $sorgu = AboneModel::where(array(
@@ -72,6 +105,15 @@ class indexController extends Controller
 
     // SILME KISMI AYARLAMASI
     public function delete(AboneModel $item){
+
+        if (myHelper::yetkiKontrol('aboneler',"silme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $sonuc = $item->delete();
 
         if ($sonuc) {
@@ -94,6 +136,14 @@ class indexController extends Controller
     // AKTIF PASIF KISMI AYARLAMASI
     public function isActiveSetter(Request $request, AboneModel $item)
     {
+        if (myHelper::yetkiKontrol('aboneler',"guncelleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = ($request->data == "true") ? 1 : 0;
         $item->update(array(
             "abone_durum" => $data
